backend/requirements.txt
fastapi
uvicorn
requests
scikit-learn
numpy
python-Levenshtein
jellyfish
pydantic
sentence-transformers
faiss-cpu
sqlalchemy
